+++ gui/controller/palettecontroller.cpp
@@ -2,16 +2,7 @@
 #include "palettecontroller.h"
 
 PaletteController::PaletteController() {
-	uint8_t num_colors = 14;
-	Colors::RGB colors[num_colors] = {Colors::BLACK};
-
-	Colors::generate_scaling_color_array(colors, &Colors::RED, &Colors::YELLOW, num_colors, true);
-	palettes_.push_back(Palette("Fire", colors, num_colors));
-
-	Colors::generate_scaling_color_array(colors, &Colors::BLUE, &Colors::GREEN, num_colors, true);
-	palettes_.push_back(Palette("Deep Sea", colors, num_colors));
-
-	palettes_.push_back(Palette("Color Wheel", Colors::COLORWHEEL, 12));
+	initialize_palettes();
 }
 
 PaletteController::Palette* PaletteController::add_palette(std::string name, Colors::RGB* colors, uint8_t num_colors) {
@@ -34,6 +25,21 @@ PaletteController::Palette* PaletteController::get_palette(std::string name) {
 
 std::vector<PaletteController::Palette> PaletteController::get_palettes() {
 	return palettes_;
+}
+
+void PaletteController::initialize_palettes() {
+	uint8_t num_colors = 14;
+	Colors::RGB colors[num_colors] = {Colors::BLACK};
+
+	palettes_.clear();
+
+	Colors::generate_scaling_color_array(colors, &Colors::RED, &Colors::YELLOW, num_colors, true);
+	palettes_.push_back(Palette("Fire", colors, num_colors));
+
+	Colors::generate_scaling_color_array(colors, &Colors::BLUE, &Colors::GREEN, num_colors, true);
+	palettes_.push_back(Palette("Deep Sea", colors, num_colors));
+
+	palettes_.push_back(Palette("Color Wheel", Colors::COLORWHEEL, 12));
 }
 
 void PaletteController::remove_palette(uint8_t index) {