+++ gui/widget/palettecontrol.h
@@ -2,6 +2,7 @@
 #define PALETTECONTROL_H
 
 #include "controller/palettecontroller.h"
+#include <QAbstractButton>
 #include <QDialog>
 
 namespace Ui {
@@ -16,20 +17,17 @@ class PaletteControl : public QDialog {
 		~PaletteControl();
 
 	private slots:
+		void on_baseColorButton_clicked();
+		void on_buttonBox_clicked(QAbstractButton *button);
 		void on_color_clicked();
 		void on_createButtonBox_accepted();
+		void on_createButtonBox_rejected();
 		void on_createPaletteButton_clicked();
 		void on_paletteComboBox_currentIndexChanged(int index);
 		void on_removeButton_clicked();
-		void set_button_color(QPushButton* button, uint8_t red, uint8_t green, uint8_t blue);
-
-		void on_createButtonBox_rejected();
-
-		void on_typeComboBox_currentIndexChanged(int index);
-
-		void on_baseColorButton_clicked();
-
 		void on_targetColorButton_clicked();
+		void on_typeComboBox_currentIndexChanged(int index);
+		void set_button_color(QPushButton* button, uint8_t red, uint8_t green, uint8_t blue);
 
 	private:
 		Ui::PaletteControl *ui;
@@ -39,6 +37,7 @@ class PaletteControl : public QDialog {
 		QColor base_color_ = Qt::black;
 		QColor target_color_ = Qt::black;
 
+		void initialize_palettes(std::string initial_palette);
 		void set_create_palette_controls_visible(bool visible);
 };