+++ src/core/colors.h
@@ -117,9 +117,9 @@ namespace PixelMaestro {
 			static RGB RAINBOW[];
 
 			static RGB generate_random_color();
-			static void generate_random_color_array(RGB* array, RGB* base_color, uint32_t num_colors, float range = 1.0);
-			static void generate_scaling_color_array(RGB* array, RGB* base_color, RGB* target_color, uint32_t num_colors, bool reverse = false);
-			static void generate_scaling_color_array(RGB* array, RGB* base_color, uint32_t num_colors, uint8_t threshold, bool reverse = false);
+			static void generate_random_color_array(RGB *array, uint8_t num_colors);
+			static void generate_scaling_color_array(RGB* array, RGB* base_color, RGB* target_color, uint8_t num_colors, bool reverse = false);
+			static void generate_scaling_color_array(RGB* array, RGB* base_color, uint8_t num_colors, uint8_t threshold, bool reverse = false);
 			static RGB mix_colors(RGB color_one, RGB color_two, MixMode mode, uint8_t alpha = 0);
 	};
 }